move arguments to package
